Post Formu ile Ürün Silme/56
Bu derste ürünü silmeyi biraz değiştiriyoruz. Yaptığımız şey, ürünü silme butonuna tıklayınca doğrudan silme durumunu değiştirip, “ürünü gerçekten silmek istiyor musunuz?” uyarısını yapacağımız bir sayfaya yönlendireceğiz ve bu sayfada verilen evet veya hayır yanıtına göre ürünü silmek olacak.
Teknik olarak yapacağımız şey, bir DeleteConfirm viewi tanımlamak, bir HttpPost Delete metodu yazmak olacak.

+++ Controllers/HomeController.cs
@@ -180,6 +180,7 @@ namespace Forms.Controllers
 
         public IActionResult Delete(int? id)
         {
+            // gönderilen ürün id'ye göre, ürün modelini entity değişkenine aktarıyoruz
             var entity = Repository.Products.FirstOrDefault(p => p.ProductId == id);
 
 
@@ -187,14 +188,29 @@ namespace Forms.Controllers
             {
                 return NotFound();
             }
-            else
+
+            return View("DeleteConfirm", entity);
+        }
+
+        [HttpPost]
+        public IActionResult Delete(int? id, int ProductId)
+        {
+
+            if (id != ProductId)
             {
-                Repository.DeleteProduct(entity);
+                return NotFound();
             }
+
+            var entity = Repository.Products.FirstOrDefault(p => p.ProductId == ProductId);
+            if (entity == null)
+            {
+                return NotFound();
+            }
+
+            Repository.DeleteProduct(entity);
 
             return RedirectToAction("Index");
         }
-
 
 
     }
